NET-1465: Database lookup transformation (#16)
https://seaseducation.atlassian.net/browse/NET-1465
---
### Problem
As an ETL developer, I want to be add or replace values in a record based on the results of a database query executed with parameters from other fields in the record.

Functionality required from previous generations:

- Option to cache results to avoid repetitive database calls when the result is already known. This should not be the default behavior.
- Option to have a zero-result lookup be valid.

New functionality that should be included:

- Ability to specify distinct event messages for zero- or multi-result lookups. This will assist in creating more meaningful message for users when there is an issue with a field.

### Solution
- Added ability to generate database connections.
- Added DbLookup transformation that has:
  - Distinct customizable event behaviors for zero or multiple results. These can be used to customize both the logged event message, as well as controlling whether they are fatal events.
  - A caching option with it own handler.

### Additional Notes
No specific database implementations exist in the core library. Each database will have its own package to minimize dependency churn in the core project.