NET-1511: Fix defect where transformation decorators are not disposing handlers (#24)
https://seaseducation.atlassian.net/browse/NET-1511
---
### Problem
While working on logging, I discovered that CSV output files generated by a dataflow could not be compressed because the file was in use.

### Solution
I tracked the issue down to a transformation decorator not disposing of its decorated handler. I was able to quickly add the missing dispose.

### Additional Notes
I had a test that should have caught this, but the test had a flaw: I had mistakenly made the unit test return `ValueTask` instead of `Task`. XUnit does not understand `ValueTask` and report all such test as passing, even when they fail.

I located all the tests returning `ValueTask` and corrected them to `Task`.