NET-1518: Add callback for database sources during reading for resource cleanup (#25)
https://seaseducation.atlassian.net/browse/NET-1518
---
### Problem
When DbSource reads data from an Oracle data reader containing blob/clob fields, these fields contain disposable types that are not properly disposed when the reader advances to the next record. 

### Solution
Added an optional callback method on provider-specific connection factory implementations that can be called when reading a record is finished so that cleanup can be performed on any record-specific resources.