NET-1527: Fix return of null record when suppressed exception thrown (#4)
https://seaseducation.atlassian.net/browse/NET-1527
---
### Problem
When a CSV file contains a duplicate header and the ThrowOnBadData option is false, a null record will be returned from the data source. This causes an exception in the dataflow pipeline rather than logging the error as intended.

### Solution
Stops reading the file when an exception is caught that is not rethrown.