NET-1589: Conversion performance improvements #30
### Problem
As a developer, I want to optimize the performance of converters to reduce the CPU cost of using the transformation. During testing of a dataflow with several conversion transformations, the process was using excessive amounts of CPU and dataflows were stalling without error.

### Solution
Better short-circuiting and avoiding the IConvertible cast when we already know the value is a string. This reduced the CPU cost dramatically and allowed conversion-heavy dataflows to execute much faster.

Also made simple delegates into static methods.