NET-1807: Additional parameters for DbLookup transformation (#33)
### Problem
As a developer, I want to pass additional static input parameters to DbLookup transformations that are not part of the dataflow record so that I can add them as query parameters without needing to add static values to every record.

### Solution
Added a parameter collection to DbLookup that will be added to the input parameters collection for the query. In the event any parameter names conflict with Input column names, the value from the record will override the static parameter value.

### Additional Notes
Also corrected the ClearField field to a property on the Conversion transformation.