NET-1616: Option to skip lines before reading CSV files #35
https://seaseducation.atlassian.net/browse/NET-1616
---
### Problem
As an ETL developer, I want an option to skip a number of rows at the beginning of a file before activating the CSV reader so that I can map fields by position when the records do not match the header row.

### Solution
Added a CSV source option that skips a specified number of lines.

### Additional Notes
The skipped lines are treated as pain text without any CSV parsing. The skipping function will count line breaks whether they are escaped or not.